wult: driver: fix C-state residency #2
We often were taking C-state snapshots incorrectly, which was leading to
C-state residencies like 1M% sometimes. This patch fixes the problem.

Instead of trying to figure out whether the C-states snapshot has to be taken
in 'after_idle()' or the interrupt handler, just take it in both of them and
figure out which one to use later, in 'wult_tracer_send_data()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>