WultRunner: simplify CSV header creation
Before this patch, we formed the CSV header by removing the fields we do not
want, then inserting the additional fields. This is hard to maintain, because
you need to touch the code every time the driver adds more fields.

With this patch, we form the CSV header from the definitions file. The
assumption is that the defs file gives us the full list of the columns that
should be saved in the datapoints CSV file.

Theres is a catch with C-states, though, and with the fact that some drivers
may not provide some of the fields. E.g., "WarmupDelay" is only provided by
'wult_igb' driver. This patch handles the situation using the 'optional' key
from the defs file.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>