wult: remove IntrDelay
Patch 'a574fcb WultRunner: use IntrOff' partially removed 'IntrDelay'. It was
meant to be done in a separate patch, though. We want to remove 'IntrDelay'
because it does not really work well, and we can see negative numbers when
measuring C1. This happens because we mix various clock sources in one
equation, and because the TSC math we do may be inaccurate to few microseconds.

So remove 'IntrDelay' for now. We may return it at some point if we figure out
how to account for the extra measurement overhead in a more accurate way.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>