WultRunner: rework datapoint formation
Before this patch, we would form the CSV datapoint by modifying the raw
datapoint in-place. This is a bit confusing at times, and not easy to follow.

With this patch, instead of modifying the raw datapoint, we create a new 'dp'
dictionary, and populate it with data, without changing the raw datapoint. This
is more clear and straight-forward way of doing it, and a bit more effecient.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>