RORawResult: do not lose CC1Derived% in older results
We renamed 'DerivedCC1%' to 'CC1Derived%', but when we generate a report form
an older result, we lose this metric, because now the code is looking for the
new name.

This patch renames the column in old results on the fly, so it won't be
dropped. This patch also increases the raw result format version from 1.1 to
1.2, since we've done significant enough changes.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>