wult: driver: add 'intr_focus' debugfs file
This patch adds the 'intr_focus' debugfs file. This file is R/W, and it
contains a boolean value.

Right now writing '1' to this file does notthing, but later it will enable the
interrupt latency focused measurements, where we'll skip taking any
measurements in 'after_idle()' and will only measure the interrupt latency in
the interrupt handler.

The default value is '0'. The mode is allowed to be changed only when the
measurements are disabled.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>