wult: add --intr-focus option
Add the '--intr-focus' option to the 'start' command to enable interrupt
latency focused measurements. In this case 'WakeLatency' is not measured, and
'IntrLatency' measurements become more precise.

This patch only adds the option, which enables the corresponding driver mode
via the debugfs file. The interrupt latency focused mode is not implemented
yet, so this option does nothing so far.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>