wult: driver: limit C-state counters
Linux kernel version 5.14 supports max. 32 fields in a synthetic trace event,
and we exceed this limit on some platforms. I've submitted a kernel patch to
increase the limit, but meanwhile, let's send less to user-space.

One large fields consumer is the C-state counters. There are many C-states
counters, and we read/send them all, but most of them are always 0.

This patch removes C-state counters that have 0 value from the list of C-state
counters that are sent to user-space. As a result, we send 6-10 items less than
we used to.

The disadvantage and risk is that if for some reason the platform supports a
C-state, but just never reached it before wult driver starts, the C-state will
be excluded.

May be later, if kernel starts supporting more fields, we can revert this
patch.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>