wult: flatten 'start_command()'
We have too many nested 'with' statements in 'start_command()', and we are
about to add another one. It becomes hard to read. Flatten it using the
'ExitStack()' context manager, which you can use to just add all the nested
contexts, which will be exited as soon as the 'ExitStack()' context is exited.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>