wult/ndl: allow for --xaxes=none and --yaxes=none
Sometimes it is useful to generate a report without a scatter-plot, for example
to make the report really small. For some reason we did not allow for this, but
with this patch, the '--xaxes' and '--yaxes' arguments now accept the 'none'
value, just like in case of '--hist' and '--chist'. With the 'none' value in
either of these options, we skip the scatter plot.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>