WultRunner: implement TSC rate calculation
Implement TSC rate calculation. In order to calculate the rate, we collect
datapoints for 10 seconds, without processiong them. But they are not lost,
they are saved. After 10 seconds, we calculate TSC rate and process all the
saved data. After this we continue processing all the new incoming raw
datapoints normally.

This is only the first step in towards the goal of doing all the TSC math in
user-space, instead of doing it in the kernel.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>