_WultDpProcess: change API
Change the API from 'process_datapoint()', which receives a raw datapoint and
returns a processed datapoint, to something a bit more complex:

* 'add_raw_datapoint(rawdp)' - add a raw datapoint to the processor. No return
  value.
* 'get_processed_datapoints()' - a generator which yields all the currently
  available processed datapoints.

This is a preparation to TSC support. We are planning to move calculations from
the driver up to userspace. For example, 'WakeLatency' will be calculated in
userspace, not in the kernel. For this, we need to know TSC rate, in order to
translate cycles to microseconds.

In order to calculate the TSC rate, we need to consume raw datapoints for
something like 1 second, and save them. Only then we can calculate the TSC
rate.

With this new API, 'WultRunner' will feed raw datapoints to
'DatapointProcessor', and will get no processed datapoints for some time, until
'DatapointProcessor' has enough data to calculate TSC rate. After these, all
those raw datapoints, saved in 'DatapointProcessor', will be processed and
yielded via 'get_processed_datapoints()'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>