wult: driver: wult_igb: provide raw timestams to userspace
We are going to get rid of TSC->nanoseconds conversion in the wult_igb driver.
Instead, the driver will provide raw TSC timestamps, and user-space will
calculate the overhead of reading NIC registers over PCIe.

This patch is the first step - just introduce raw TSC couners. We'll remove the
TSC->nanoseconds conversion in a later patch.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>