wult: driver: read SMI/NMI counters at the right time
We need to take SMI/NMI counters first before, and then after everything else.
The "before" part is OK, but the "after" part was not - we took the TSC
snapshot after reading SMI/NMI counters. But there can be an SMI/NMI right
beofore reading the TSC snapshot. This patch fixes this potential issue.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>