wult: driver: introduce a mechanism for time adjustments
The 'wult_igb' driver talks to a NIC behind potentially slow PCI bus, and this
makes fast idle states measurements very inaccurate. We have a mechanism for
adjusting for some of these inaccuracies, but this mechanism does not work
correctly today.

This patch is the first step towards improving and fixing that mechanism. We
add a 'time_adj' argument to the 'get_time()' functions, so that driver can
return the time in whatever units, and the adjustment for the time in TSC
cycles. Only 'wult_igb' will use it. We'll then propagate the adjustments to
user-space and apply them in user-space.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>