wult: driver: add CPU cache dirtying feature
Deep C-states like C6 need to flush CPU cache, which will be longer when there
are dirty data in it. Therefore, sometimes we want to measure exit latency for
the "worst case", when entire CPU cache contains only "dirty" cache lines.

This patch adds a 'drity_cache_buf_size' debugfs file. By default it contains
zero.

Users may write the amount of byts that have to be written to before requesting
a C-state. For example, writing 2M would make wult allocate a 2M buffer, and
fill it with zeroes before requesting a C-state.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>