_HTMLReportBase: validate hover text columns
Before this patch, if the hover text columns include a column like
'WarmupDelayCyc' (valid column name, it is present in the trace data, but we do
not have it in the definitions), wult crashes with "KeyError:
'WarmupDelayCyc'".

This patch adds a piece of validation code that will just drop all columns that
are not in defs and print a warning. We already do it for 'xaxes' and the like,
but not for hover text columns, and this patch extends the validation to hover
text columns. With this patch, we have:

wult: warning: dropping column 'WarmupDelayCyc' from hover text because it is
      not present in the definitions file at 'defs/wult.yml'

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>