_HTMLReportBase: change css/style.css to copy to output dir
Symbolic linking to CSS files can cause the styles to fail to load. The
Bootstrap assets already have a fallback option to use the CDN and
remotely load the assets. This fallback is not available for wult's
style.css therefore now we copy this file to the output directory no
matter what 'self.relocatable' is set to.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>