ToolsCommon: adjust to 'cpugeom' 1st level keys change
The 'cpugeom' dictionary structure has changed, and now the first lelvel key
names are the same 'level' (scope) names, like "core" or "cpu". Everything
related to cores, is in 'cpugeom["core"]', everything related to CPUs is in
'cpugeom["CPU"]', and so on. There will be more 'cpugeom' dictionary
improvements in subsequent patches.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>