RORawResult: move summary table logic to 'DFSummary'
In future patches, statistics tabs such as 'AC Power' will be introduced
to the HTML report and include their own summary tables. This patch is
the first of a few which will aim to prevent code duplication by
extracting the class-independent logic from summary table-related
functions so that it can be reused by the upcoming statistics tabs.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>