_Plot: remove microseconds hack from '_Plot' base class
This patch modifies '_Plot' so that the baseclass no longer contains a
special and inconsistent clause for the unit argument value
'microseconds'.

To handle the fact that some metrics have a unit definition of 'us' or
'microseconds', this patch catches this situation in '_PlotsBuilder' and
documents the reason for the clause.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>