RORawResult: fix row selector for older 'numexpr'
Before this patch, the following command fails on some systems:

  $ wult report --rsel 'PC6% > 0.0'

ValueError: data type must provide an itemsize

This appears to be a bug in some 'numexpr' versions. For example, this happens
on Fedora 33 when the 'python3-numexpr-2.7.1-4' package is installed.

This patch adds a work-around - we fall-back to using the "python" engine
instead of the "numexpr" engine if the above error happens.

Reported-by: David Fang <david.y.fang@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>