html: templates: remove Bootstrap
The main reason for introducing 'Bootstrap' was to help with adding tabs
to wult reports. Tabs have since been replaced with a better solution
provided by 'shoelace.style'.

This means that we can now remove the dependency on 'Bootstrap', the
minified files in the repo and the logic around looking for a local copy
of 'Bootstrap' and falling back on the CDN.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>