_ReportBase: prevent stats tab generation if a report has no stats
Before this patch, stats tab builders were unaware of reports which did
not contain any statistics. This means that they were unable to stop the
tab generating. This patch changes it so that 'ReportBase' passes all of
the reports to stats tab builders along with whether or not they have a
stats directory.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>