_ACPowerTab: reduce column name usage
To simplify function signatures and data structures, this patch limits
column name usage to only the parsing of raw AC Power files.

Specifically, statistics tabs should not need to maintain a mapping from
column names to metrics. That is, columns should be renamed to match the
names of metrics where possible.

This patch is the first in a series in which each patch does the same
thing on a different class or base class.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>