RORawResult: fix 'wult calc --rsel'
The CSV file we load contains more columns than we have metrics in the 'defs'
file. E.g., the 'SilentTimeRaw' CSV file column is not described in the defs
file.

This file fixes a bug which is related to accessing defs for columns which are
not present in defs, such as 'SilentTimeRaw'. Here is an example of a failure:

$ wult calc --rsel 'PC8% > 0' results/
Loading test result 'result/datapoints.csv'.
Traceback (most recent call last):
  File "wult", line 19, in <module>
    sys.exit(main())
  File "wultlibs/_Wult.py", line 592, in main
    args.func(args)
  File "wultlibs/ToolsCommon.py", line 434, in calc_command
    res.calc_smrys(funcnames=funcnames, all_funcs=all_funcs)
  File "rawresultlibs/RORawResult.py", line 157, in calc_smrys
    colnames = self.get_numeric_colnames(colnames=all_colnames)
  File "wultlibs/rawresultlibs/RORawResult.py", line 53, in get_numeric_colnames
    if self.defs.info[colname]["type"] in ("int", "float"):
KeyError: 'SilentTimeRaw'

The fix is to exclude columns, such as 'SilentTimeRaw', from self.columns.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>