_MetricTabBuilder: use 'fsname' for paths of generated assets
Fixes the following command: `wult report results/ -y='CC0%'`.
Previously, using the `-y` option with any C-state metric would cause a
broken report to be generated. This is because the assets for the tab
containing the C-state metric would be located within a directory with a
"%" in the name. This made those assets unreachable to browsers.

This patch fixes the issue by ensuring that metric tabs use 'fsname' to
dictate where assets are created.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>