js: smry-tbl: remove duplication of 'getWidth'
'getWidth' (defined in 'report-table') was overridden in 'smry-tbl'. All
tables will switch to not using table dictionaries but parsing table
text files. This means that the version of 'getWidth' in 'smry-tbl' will
become the way used by all report tables. Therefore this patch updates
the definition in the parent class and removes the definition in the
child class.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>