_ReportBase: reduce output of stats tab generation errors
This patch alters the error handling of stats tab generation so that if
an error is raised during tab generation, a standard message is
presented to the user. If the user uses the '-d' debug option then they
will see the details of the error.

This change also makes this more consistent with how errors are handled
during the initialisation of statistics tab generators.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>