WultDefs: do not rely on 'populate_cstates()'
Stop using the 'populate_cstates()' method for building wult definitions.
Instead, mangle the definitions in the constructor, so that users do not have
to call 'populate_cstates()'. The 'WultDefs' class just knows how to mangle its
definitions, and users do not need to worry about that.

Do not remove the 'populate_cstates()' method so far, because it is still used
by '_TurbostatL2TabBuilderBase'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>