_TurbostatL2TabBuilderBase: add '_defs' property
Defs are used in several parts of the 'TurbostatL2TabBuilderBase' class
but there is no need to instantiate them multiple times. For this
reason, this patch introduces the '_defs' property which will be
instantiated once with 'TurbostatDefs'.

In following patches, other 'defs' references in the class will be
changed to use this property.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>