Deploy: cleanup terminology and more
This is a big and messy commit, which I decided not to split to save time.

Them main thing this commit does is it fixes terminology. Before, we used word
"helpers" to refer both any helper program, and non-pytohn helper programs. Now
we use word "simple helper" for a non-python stand-alone helper probram, and
word "helper" for any helper program.

While doing this change, I searched for word "helper" and reviewed everything
aroung the found lines. So there are some more little cleanups in this patch.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>