Deploy: add BPF helpers support
We are working on using BPF programs for generating interrupts. The work is not
finished yet, but it is helpful to have general BPF helpers dployment
infrastructure merge early.

This patch adds a notion of a "bpf helper" to 'Deploy'. If a tool has one, the
helper will be deployed as well. BPF helpers require 'clang', 'bpftool', and
'libbpf.a' library. This patch adds the corresponding checks.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>