EventsProvider: add factory function
Introduce the 'EventsProvider' factory function. It will create and return the
events provider object suitable for the device it was given.

At this point we support only devices controlled by wult drivers. But we are
working on other devices support, which will have a different events provider
class implementation. This patch is a preparation.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>