_MetricDTabBuilder: rename 'self.tabname' to 'self._tabmetric'
'tabname' is generic and does not make it clear that the variable
contains a metric name despite the fact that many of it's references
require that it is a metric name.

This patch amends this and makes the property private. The property
should not be public since a caller changing the property would most
likely break the generation of the tab.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>