_IPMITabBuilder: use metric names for summary tables and plot axes
If a report viewer wants to screenshot the plot or the summary table, it
might not make a lot of sense to others what the IPMI column represents.
Therefore after this patch, wult uses the metric names instead of the
column names for these labels.

For example, a summary table and plot labelled "Fan Speed" is much
clearer than "Fan2A".

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>