_Ndl: create device object
This patch is a small step towards making 'WultRunner' and 'NdlRunner' be
more similar.

'WultRunner' class accepts the device object, while 'NdlRunner' accepts the
network interface object.

This patch is a preparation for making 'NdlRunner' also accept a device object.
In this patch we just create the device object. The next patch will pass it
down to 'NdlRunner'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>