_PlotsBuilder: add error handling for results with different metrics
Not all results contain data for the same metrics. If one result does
not contain data for a given metric, it will cause an error when
building the plot as it will look for data which doesn't exist. This
patch adds error handling which raises an appropriate error to clarify
what has happened.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>