WultRunner/NdlRunner: print correct progress line in case of errors
Before this patch, errors caused the progress line to print "0" datapoints
collected. This is because '_collected_cnt' was never updated when '_collect()'
raised an exception.

This patch makes the 'ProgressLine' object save the last printed datapoints
count and latency, so we can re-print them if needed. This also makes
'self._max_latency' variable unneeded.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>