_RawDataProvider: fix drivers unloading
This patch fixes the following error:

wult: error: ran the following command 'mhl0', but it failed with exit code 1:
rmmod wult
rmmod: ERROR: Module wult is in use by: wult_igb

This happens when we start wult with 'tdt' method, but there are already
(leftover) drivers for 'wult_igb' method loaded.

The fix is to always unload all drivers on the initialization.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>