_WultrunnerRawDataProvider: add BPF raw data provider implementation
The 'hrt-bpf' method does not use kernel drivers, but it used eBPF program. So
the way we read raw datapoints is different: we do not read them from the
ftrace  buffer, but we read them from the 'wultrunner' process.

This patch implements the '_WultrunnerRawDataProvider' class which handles the
BPF case. The '_DrvRawDataProvider' handles the driver case.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>