Devices: add tool-specific I210 classes
The Intel I210 device can be used with both wult and ndl. In case of wult, we
only need a driver, in case of ndl, we need a (different) driver and a helper
(ndlrunner).

Introduce tool-specific classes for Intel I210: _WultIntelI210 and
_NdlIntelI210. They are mostly identical, but have few differences.
  1. Different driver names.
  2. The ndl class also provides the helper tool name.
  3. The ndl class also verifies that the network interface exists.

This change is mostly a clean-up at this point, but it also provides a
structure for future wult eBPF support, where some wult devices will also come
with a helper tool.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>