Deploy: make 'add_deploy_cmdline_args()' independent
Turn 'add_deploy_cmdline_args()' from being a method to an independent
function. This function is uses at veryearly stages, and creating a 'Deploy'
object at these stages is problematic, because we need to know and connect to
the SUT from 'Deploy.__init__()'.

So turn it into an independent function which does not need to know about the
SUT and does not try to connect to it.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>