_TotalsL2TabBuilder: check if extra tabs are in 'common_metrics'
This class adds new data tabs to the tab hierarchy in
'_get_tab_hierarchy()'. This patch amends the method so that it checks
if all results have data for the extra metrics before adding them to the
tab hierarchy.

This will minimise the number of warnings the user sees because of
skipped tabs.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>