Deploy: fix deploying pyhelpers
This patch fixes the following issue.

1. User clones 'wult'.
2. User runs './wult deploy -H <sutname>' from the cloned repo.

In this case python helpers deployment will fail like this:

wult: error: program 'stats-collect' was not found in $PATH

The reason is that the deployment code incorrectly assumes that python helpers
are always available in 'PATH', because they were installed via setup.py. But
this is incorrect.

Fix the problem by looking in both 'PATH' and source code directory.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>