Deploy: do not require kernel sources at init time
This patch fixes the following issue.

$ wult start hrt
wult: error: kernel sources directory '/lib/modules/5.18.0-rc1+/build' does not exist on host '192.168.1.116'

This happens because we are trying to ensure kernel sources exist on the SUT in
'Deploy.__init__()', even though they do not have to exist when we do not
actually use 'Deploy.deploy()'.

This patch moves the kernel sources initialization part from the constructor to
'Deploy.deploy()'. This fixes the problem.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>