_MetricDTabBuilder: fix `wult report --size large` error
The "ReqCState" metric does not have any unit but '_add_scatter()'
assumed all scatter metrics must have short units. This caused `wult
report --size large` to crash while generating the diagrams for
"ReqCState".

After this point, there is no assumption that the metric
has a short unit so it is unnecessary to make it a requirement in
'_add_scatter()'. Therefore this patch removes that requirement by
setting a default "None" value if a short unit is not available.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>