_DTabBuilder: build 'smrytbl' without requiring unit or description
If a metric definition does not contain a unit or a description,
'DTabBuilder' will fail when it tries to add it to the summary table.
This is unnecessary as the summary table does not require either of
these things.

This patch removes this requirement and allows a summary table to be
generated without a description or unit.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>