Deploy: cleanup and improve terminology
It is hard to continue adding features to 'Deploy.py' because we don't have
enough terms. This patch introduces a new term - installable, and adds a piece
of documentation on how an installable is different from a deployable.

Installable is basically a driver or any type of helper.

Add some documentaion, do renames and adjust comments.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>