Deploy: simplify CM temporary directory logic
There are certain conditions where the CM temporary directory is not needed, so
we do not create it. But there are multiple of them, and it is prone to errors
to get them all correctly in an "if" statement.

It is easier to just create it on demand. Introduce the "_get_ctmpdir()" helper
for this - it can be called whenever a CM temporary directory is required.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>