Improve python the way we run python scripts
Sometimes we run small python scripts, and we rely on having 'python' in
'$PATH'. This works most of the time, but there are setups out there which do
not have 'python' in $PATH.

Use 'pman.get_python_path()' which tries to find the python interpreter by
checking standard possibile locations.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>