wult: drivers: ndl: Add missed header
The rule of thumb is to include what we are direct users of, excluding some,
which are guaranteed to be included by others. The types.h is not guaranteed
to be included by others, while in practice many include it. To be on the
strict side of the rule add missed types.h.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>