_DTabBuilder: reduce data before generating hover text for scatter plots
Before this patch, the plots would contain hover text for points which
did not exist in the plot. This caused plot files to grow greatly in
size.

This was caused by commit 62ce5df9020d019722a03d87d0ab2f66374ac59d.
Which changed the order of reducing the 'pandas.DataFrame' and
generating hovertext.

This is because hover text was generated for all data points in a
'pandas.DataFrame' before it was reduced. This patch amends this so that
hovertext is generated after the 'pandas.DataFrame' is reduced. This
greatly reduces the size of plots and therefore also the load times of
wult HTML reports.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>