_DTabBuilder: rework 'add_smrytbl()' to increase smry func customisation
This patch is part of the transition away from visualisation being
dictated by defs files. This patch allows the caller to specify which
summary functions should be added to the summary table for each metric.

Before this patch, the caller could only specify one set of functions
which would be added for all applicable metrics.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>