js: file-preview: set diff window to have a fixed size
The default behaviour of the HTML diff is to take up a small amount of
space. This meant that the diff tabs were usually smaller than the
actual file preview tabs.

This patch fixes the size of the diff tab so that it will fit as much of
the diff in as possible without exceeding the maximum height.

The fixed height does mean that sometimes the diff tab will contain
excessive padding.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>