Rework C-states interrupt order detection
This patch changes how we detect interrupt order for C-state, and it is
necessary to make wult work in Linux kernel 5.19 and up. See the following
Linux kernel commit for more information:

32d4fd5751ea cpuidle,intel_idle: Fix CPUIDLE_FLAG_IRQ_ENABLE

Before this patch wult drivers tried to precisely detect the interrupt order
for every event and provide the "IntrOff" flag to user-space. But now it is
much harder to do in the driver, at least without introducing ugly hacks.

With this patch, we no longer try to detect interrupt order in the driver. We
just take the time-stamps in the interrupt handler and 'after_idle()', and send
them to user-space.

User-space figures out interrupt order for every C-sate instead.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>