_IPMITabBuilder: fix invalid HTML report generation
Before this patch, if two or more reports have IPMI statistics but do
not have any common IPMI metrics, then any report generated with them
will generate an invalid HTML report.

When the invalid report is opened, only the title and intro table will
be shown. In addition, the following error will be printed to the
browser console:

> "Uncaught (in promise) TypeError: Cannot read properties of null
  (reading 'name')"

If there are no common IPMI metrics between the reports, an IPMI tab
can't be generated. The 'IPMITabBuilder' class will return a 'None'
value rather than a valid tab. This is the cause of an invalid report.

This patch changes 'IPMITabBuilder' to raise an exception if this
condition happens. This means that the exception will be handled and the
IPMI tab will be skipped rather than generating an invalid report.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>