_TabBuilderBase: change '_build_ctab()' to raise 'Error'
Before this patch, '_build_ctab()' simply returned 'None' if the
container tab could not be built. This means that each caller should
check the returned value and raise an error if it was 'None'.

To skip this duplication for all callers, this patch changes this base
class to raise the error itself.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>