_MetricDTabBuilder: change 'get_base_si_unit()' to accept an 'mdef'
This helper method originally just accepted a unit as a string literal
and would return the appropriate base SI unit as a string literal too.

Some defs were being mangled so that the "short_unit" and "unit" fields
were inconsistent with each other. This patch changes the function to
accept an entire metric definitions dictionary. Then it adjusts all of
the unit fields to keep them consistent.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>