_Plot: use 'plotly' hover templates instead of pre-defined hovertext
By using hover templates, plot classes can use the same formatting as
they use on plot axes. This means that they can automatically scale the
datapoints so that they are readable in hovertext and add SI prefixes if
necessary.

In turn, this partially solves the issue of some metrics showing '0' in
hovertext when they have already been scaled. The bug will be completely
fixed when the units are also adjusted accordingly.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>