_ReportBase: scale all results before generating results tabs
By scaling all of the results at the beginning of the tab generation
process, it will allow us to remove scaling in other parts of report
generation in following patches.

Because all of the 'defs' are updated accordingly when they are scaled,
this patch also fixes the bug where some metrics showed 0 in hovertext.
Previously, the numbers were fixed in a previous patch but the units
were still incorrect when a metric had previously been scaled.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>