_TabBuilderBase: create 'outdir' after parsing raw stats files
Before this patch, if there was a problem during raw statistics file
discovery (e.g. if one of the results did not include any statistics
file), then the tab wouldn't be generated but the 'outdir' would still
be created.

This patch reverses the order of those two operations. This means that
the 'outdir' will only be generated if raw statistic files have been
found and parsed for each result.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>