_DTabBuilder: remove a check for skipping plot generation
The check being removed by this patch will prevent a plot from being
generated if any of the results contain no non-zero datapoints. This
meant that if one result contained interesting datapoints but another
had all zero-datapoints, the plot would be skipped.

So that the report still contains a plot visualising the interesting
datapoints, this patch removes the check.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>