_DTabBuilder: fix constant value check in '_skip_metric_plot()'
The goal of the check adjusted by this patch was originally to check if
every datapoint in a plot was the same for each result. However, before
this patch, a plot would be skipped if each result only contained
datapoints of a single value, even if this value was different between
results. After this patch, a plot will only be skipped if every
datapoint in all of the results has the same value.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>