_DTabBuilderBase: prevent "SysInfo" tab from failing if 'outdir' exists
Before this patch, an 'ErrorExists' would be raised if 'wult report' was
used with output directory which alrady contained any "SysInfo" files.
This error caused the tabs to not be generated.

This patch makes it so that "SysInfo" tabs will catch the 'ErrorExists'
and keep the original "SysInfo" file in place. This fixes the issue
above and will allow "SysInfo" tabs to be generated when 'wult report'
is writing to a directory already containing a report.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>