_WultRawDataProvider: fix ldist changing bug
Fix a bug with the following symptom:

$ wult start --ldist 4001,10000 tdt
wult: error: can't to change launch distance range
failed to open '/sys/kernel/debug/wult/ldist_from_nsec' and write 4001000 to it:
        method 'close()' failed for /sys/kernel/debug/wult/ldist_from_nsec: [Errno 22] Invalid argument

The problem was that we first tried to set 'ldist_from_nsec', which was greater
than currently set 'ldist_from_nsec=4000', and we got that failure because the
of the driver check.

Fix this by changing 'ldist_to_nsec' before 'ldist_from_nsec'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>