_WultDpProcess: fix 'IntrLatencyRaw' crash
This patch fixes a crash with the following symptoms:

File "_WORawResultBase.py", line 140, in add_csv_row
     self.csv.add_row(self._get_csv_row(dp))
File "_WORawResultBase.py", line 118, in _get_csv_row
     val = dp[key]
KeyError: 'IntrLatencyRaw'

The crash is rare, but may happen.

The problem was happening because the first datapoint had 'IntrLatencyRaw', and
it became part of the CSV header. Then one of the subsequent datapoints did not
have it, and we got a crash.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>