setup.py: fix wult report regression
In v1.10.17, local viewing scripts were added to each report. However
the scripts were not added as data files in 'setup.py'. This meant that
they were not installed properly. Then 'wult report' would complain
because it could not find the scripts.

This patch fixes that regression by adding the scripts as data files in
'setup.py'.

This regression was caused by these commits:
- bf5debe80479232445cbf4f5bdefdd18a02202ef
- 219e4eec5e02d08e267f47ae79916e9ebfb9cb16

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>