wult/ndl: switch to per-project environment variables
We use environment variables for specifying wult/ndl data and helpers path. And
these variables are per-tool, so there are 4 of them:

NDL_DATA_PATH
NDL_HELPERSPATH
WULT_DATA_PATH
WULT_HELPERSPATH

And if we add a 3rd tool, there will be 6 of them. This is inconvenient and
this commit turns the environment variables from per-tool to per-project, so
there will only be two "WULT_*" variables.

The 'find_app_data()' method becomes 'find_project_data()'. All callers are
adjusted, and a couple of improvements are made at the same time (the 'descr'
argument added, an unnecessary 'self._projname' is removed).

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>