_WultCommon: fix metrics with "%" in name being ignored in report params
'get_axes()' is used to retrieve metrics to include as various
diagram-related arguments of 'WultReport'. This function escapes the "%"
character so that the text is compatible with 'argparse' helptext.

'RORawResult.find_metrics()' is used to convert 'WultReport' arguments
to metric names in '_ReportBase._init_metrics()'. However, '_ReportBase'
does not expect the "%" to have been escaped. This meant that metrics
such as "CC0%" could not be used in any 'WultReport' argument.

This patch fixes the above issue by adding the 'escape_percent'
parameter to 'get_axes()'. This allows the caller to choose whether "%"
characters should be escaped or not.

Signed-off-by: Adam Hawley <adam.james.hawley@intel.com>