ndl: workaround for annoying paramiko traceback
Often we see the following traceback when running the 'ndl' tool:

Traceback (most recent call last):
  File "/usr/lib/python3.10/site-packages/paramiko/file.py", line 66, in __del__
    self.close()
  File "/usr/lib/python3.10/site-packages/paramiko/file.py", line 84, in close
    self.flush()
  File "/usr/lib/python3.10/site-packages/paramiko/file.py", line 92, in flush
    self._write_all(self._wbuffer.getvalue())
ValueError: I/O operation on closed file.

This patch is a workaround. The traceback is caused by running 'flush()' on
stdin. I am not 100% sure, but I think this is a paramiko bug.

I just added a workaround which seems to work well.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>